requirements.txt
slack_sdk>=3.27.0
openai>=1.40.0
pyyaml>=6.0.1
